split chat_manager file contents into manager, model, message files
